# openapi/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/kyc/receivers/{receiverId}/bank-accounts:
    post:
      operationId: BankAccountsController_create_v1
      parameters:
        - name: receiverId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBankAccountDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccountEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Add a fiat bank account for a receiver
      tags: &a13
        - kyc
    get:
      operationId: BankAccountsController_findAll_v1
      parameters:
        - name: receiverId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccountListEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: List a receiver's bank accounts
      tags: *a13
  /v1/kyc/receivers/{receiverId}/bank-accounts/{id}:
    delete:
      operationId: BankAccountsController_remove_v1
      parameters:
        - name: receiverId
          required: true
          in: path
          schema:
            type: string
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Delete a bank account
      tags: *a13
...
  /v1/customers:
    post:
      operationId: CustomersController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Create a customer
      tags: &a16
        - customers
    get:
      operationId: CustomersController_findAll_v1
      parameters:
        - name: take
          required: false
          in: query
          schema:
            maximum: 100
            default: 100
            type: number
        - name: skip
          required: false
          in: query
          schema:
            default: 0
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: List the consumer's customers (with payment stats)
      tags: *a16
  /v1/customers/{id}:
    get:
      operationId: CustomersController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Get a customer by id
      tags: *a16
    patch:
      operationId: CustomersController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Update a customer
      tags: *a16
    delete:
      operationId: CustomersController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDeletedEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Delete a customer
      tags: *a16
...
info:
  title: Cosmos Pay — Payments API
  description: "Payments microservice (Stellar payment intents). All endpoints
    require traffic to arrive through the APISIX gateway: a valid
    `X-Gateway-Secret` header plus an authenticated consumer
    (`X-Consumer-Username`). Paths already include the version (`/v1/...`)."
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    gateway-secret:
      type: apiKey
      in: header
      name: X-Gateway-Secret
    consumer:
      type: apiKey
      in: header
      name: X-Consumer-Username
  schemas:
...
    CreateBankAccountDto:
      type: object
      properties:
        type:
          type: string
          enum:
            - wire
            - ach
            - rtp
            - pix
            - pix_safe
            - ted
            - spei_bitso
            - transfers_bitso
            - ach_cop_bitso
            - international_swift
            - sepa
          example: ach
        name:
          type: string
          example: Acme payouts — USD
        account_class:
          type: string
          enum:
            - individual
            - business
          example: individual
        account_type:
          type: string
          enum:
            - checking
            - saving
          example: checking
        beneficiary_name:
          type: string
          example: Jane Doe
        recipient_relationship:
          type: string
          example: first_party
        account_number:
          type: string
          example: "0123456789"
        routing_number:
          type: string
          example: "021000021"
        business_industry:
          type: string
          example: "541511"
          description: NAICS code; required for business-class US/SWIFT accounts.
        phone_number:
          type: string
          example: "+12025550123"
        tax_id:
          type: string
          example: "123456789"
        pix_key:
          type: string
          example: jane@acme.com
        pix_safe_bank_code:
          type: string
        pix_safe_branch_code:
          type: string
        pix_safe_cpf_cnpj:
          type: string
        ted_bank_code:
          type: string
        ted_branch_code:
          type: string
        ted_cpf_cnpj:
          type: string
        transfers_account:
          type: string
          example: "0000003100010000000001"
        transfers_type:
          type: string
          example: CBU
          description: CVU | CBU | ALIAS
        spei_clabe:
          type: string
          example: "012345678901234567"
        spei_institution_code:
          type: string
        spei_protocol:
          type: string
        ach_cop_beneficiary_first_name:
          type: string
        ach_cop_beneficiary_last_name:
          type: string
        ach_cop_document_id:
          type: string
        ach_cop_document_type:
          type: string
        ach_cop_email:
          type: string
        ach_cop_bank_code:
          type: string
        ach_cop_bank_account:
          type: string
        swift_account_holder_name:
          type: string
        swift_account_number_iban:
          type: string
        swift_code_bic:
          type: string
        swift_bank_name:
          type: string
        swift_bank_address_line_1:
          type: string
        swift_bank_city:
          type: string
        swift_bank_state_province_region:
          type: string
        swift_bank_country:
          type: string
        swift_bank_postal_code:
          type: string
        swift_beneficiary_address_line_1:
          type: string
        swift_beneficiary_city:
          type: string
        swift_beneficiary_state_province_region:
          type: string
        swift_beneficiary_country:
          type: string
        swift_beneficiary_postal_code:
          type: string
        swift_bank_address_line_2:
          type: string
        swift_beneficiary_address_line_2:
          type: string
        swift_payment_code:
          type: string
          description: SWIFT payment/purpose code.
        swift_intermediary_bank_swift_code_bic:
          type: string
        swift_intermediary_bank_account_number_iban:
          type: string
        swift_intermediary_bank_name:
          type: string
        swift_intermediary_bank_country:
          type: string
        sepa_iban:
          type: string
        sepa_beneficiary_bic:
          type: string
        sepa_beneficiary_legal_name:
          type: string
        sepa_beneficiary_address_line_1:
          type: string
        sepa_beneficiary_city:
          type: string
        sepa_beneficiary_postal_code:
          type: string
        sepa_beneficiary_country:
          type: string
        sepa_beneficiary_address_line_2:
          type: string
        sepa_beneficiary_state_province_region:
          type: string
        address_line_1:
          type: string
        address_line_2:
          type: string
        city:
          type: string
        state_province_region:
          type: string
        country:
          type: string
        postal_code:
          type: string
      required:
        - type
        - name
    BankAccountEntity:
      type: object
      properties:
        id:
          type: string
          example: clz9xbank00001
        blindpayId:
          type: string
          example: ba_000000000000
        rail:
          type: object
          example: ach
        name:
          type: object
          example: Acme payouts — USD
        country:
          type: object
          example: US
        createdAt:
          format: date-time
          type: string
          example: 2026-06-28T12:00:00.000Z
      required:
        - id
        - blindpayId
        - createdAt
    BankAccountListEntity:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/BankAccountEntity"
        total:
          type: number
          example: 1
      required:
        - data
        - total
...
    CreateCustomerDto:
      type: object
      properties:
        name:
          type: string
          example: Acme Inc.
        alias:
          type: string
          example: Acme — billing
        note:
          type: string
          example: VIP customer, net-30 terms.
        email:
          type: string
          example: billing@acme.com
        account:
          type: string
          example: GCALNQQBXAPZ2WIRSDDBMSTAKCUH5SG6U76YBFLQLIXJTF7FE5AX7AOO
          description: Optional Stellar account to associate the customer with.
        reference:
          type: string
          example: cust_001
      required:
        - name
    CustomerEntity:
      type: object
      properties:
        id:
          type: string
          example: clz9xcust00001
        name:
          type: string
          example: Ada Lovelace
        alias:
          type: object
          nullable: true
          example: Acme — billing
        email:
          type: object
          nullable: true
          example: ada@example.com
        account:
          type: object
          nullable: true
          description: Stellar account associated with this customer, if any.
          example: GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
        note:
          type: object
          nullable: true
          example: Prefers wire settlement
        reference:
          type: object
          nullable: true
          example: ext_4821
        createdAt:
          format: date-time
          type: string
          example: 2026-06-21T12:34:56.000Z
        updatedAt:
          format: date-time
          type: string
          example: 2026-06-21T12:34:57.000Z
      required:
        - id
        - name
        - createdAt
        - updatedAt
    CustomerWithStatsEntity:
      type: object
      properties:
        id:
          type: string
          example: clz9xcust00001
        name:
          type: string
          example: Ada Lovelace
        alias:
          type: object
          nullable: true
          example: Acme — billing
        email:
          type: object
          nullable: true
          example: ada@example.com
        account:
          type: object
          nullable: true
          description: Stellar account associated with this customer, if any.
          example: GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
        note:
          type: object
          nullable: true
          example: Prefers wire settlement
        reference:
          type: object
          nullable: true
          example: ext_4821
        createdAt:
          format: date-time
          type: string
          example: 2026-06-21T12:34:56.000Z
        updatedAt:
          format: date-time
          type: string
          example: 2026-06-21T12:34:57.000Z
        payments:
          type: number
          description: Payment intents attributed to this customer.
          example: 12
        succeeded:
          type: number
          description: How many of those succeeded.
          example: 11
        total:
          type: string
          description: Gross settled volume, an exact decimal string.
          example: "1042.5"
      required:
        - id
        - name
        - createdAt
        - updatedAt
        - payments
        - succeeded
        - total
    CustomerListEntity:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CustomerWithStatsEntity"
        total:
          type: number
          description: Total matching rows — not the page length.
          example: 137
        take:
          type: number
          example: 100
        skip:
          type: number
          example: 0
      required:
        - data
        - total
        - take
        - skip
    UpdateCustomerDto:
      type: object
      properties:
        name:
          type: string
          example: Acme Inc.
        alias:
          type: string
          example: Acme — billing
        note:
          type: string
          example: VIP customer, net-30 terms.
        email:
          type: string
          example: billing@acme.com
        account:
          type: string
          example: GCALNQQBXAPZ2WIRSDDBMSTAKCUH5SG6U76YBFLQLIXJTF7FE5AX7AOO
          description: Optional Stellar account to associate the customer with.
        reference:
          type: string
          example: cust_001
    CustomerDeletedEntity:
      type: object
      properties:
        id:
          type: string
          example: clz9xcust00001
        deleted:
          type: boolean
          example: true
      required:
        - id
        - deleted
...
security:
  - gateway-secret: []
  - consumer: []